requirements.txt
# 依赖管理
volcenginesdkarkruntime>=1.0.0
pyautogui>=0.9.54
Pillow>=10.0.0
pyperclip>=1.8.2
python-dotenv>=1.0.0
